Broad and brief overview of user levels extension.
I'm calling this extension "base" as it doesn't do much on it's own, and
is more or less just a draft of the documentation of a much larger
idea-to-come. The extension itself just sets a user level, but doesn't
work with it in any way. This initial commit leaves a lot of TODO's,
but should give a good frame of reference for the undertaking at hand.

Uses the userprefs class so is under the preferences extension branch.