The first of many... this is more of a demo:
First, let's make sure everyone's on the same page here.
This is a demo of how the other extensions could be implemented.
Feedback? If all looks good with this format, or if anyone has a
better idea and we decide on that one, we'll use the end changes to
this extension as a base template to apply to all extensions.