It's Shimmie without Shimmie!
As Shish wrote in the Shimmie documentation, Shimmie runs on SCore, and if
you remove the imageboard functions you can extend it to do whatever else
you want. Well, I say let's do it! I'm going to modify the comment extension
later to make it more general, without an "image_id" but instead a "uniq_id"
which is text, so it can be used in anything, and I'm going to modify other
extensions to make them non-imageboard-style.